requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
attrs==25.3.0
certifi==2025.6.15
cffi==1.17.1
h11==0.16.0
idna==3.10
outcome==1.3.0.post0
pycparser==2.22
pysocks==1.7.1
python-dotenv==1.1.1
selenium==4.34.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.30.0
trio-websocket==0.12.2
typing-extensions==4.14.1
urllib3==2.4.0
websocket-client==1.8.0
wsproto==1.2.0